Make ATLAS really optional
ATLAS was not marked as required in the CMakeLists files, but in
practice, it was used in the code unconditionally.
With the updated boost-bindings, we can choose not to use ATLAS.
A new cmake variable WITHOUT_ATLAS, if set to true, will
disable the use of ATLAS. With the reference BLAS, CBLAS and LAPACK, it
seems to be working (all tests in Kernel pass and in Numerics only
test-LCP_LATIN-lcp_ortiz and test-LCP_LATIN_W-lcp_ortiz are failing).
With openblas, it's a completly different story.

TODO : make ATLAS optional in the Front-End and IO

git-svn-id: svn+ssh://gforge/svn/siconos/trunk@3331 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5