Mega commit to introduce new EventsManager
The EventsManager has been rewritten. Now all the hard work is done when
inserting the event. Processing is now really easy, just pop from the
list the event. It has also been simplified a lot. Tests in the SMC
contexts are good, in the sense that no time is wasted sorting again and
again the events.

An old feature has been updated. Now the Simulation stops when there are
no more Events to process. This seems to have been the way Siconos
worked a few years ago. A few bugs where discovered thanks to this
change.

Son's examples should fail as before and ObserverLCS needs some fixing