Improve sliding mode controllers
Add a new controller based on a new scheme that design the smc
controller on the sampled plant (LinearSMCimproved)

The LinearSMC controller can now use a theta method on the equivalent
part of the control. The integrator is also ZOH

The ZeroOrderHold has been extended and improved. Now we can also
compute an integral of a matrix exponential. We now also handle better
the cases when A and/or B are constant

Add new examples of an unstable 4D system to illustrate the
discretization error on the equivalent part