[Algebra] clean up the interface (ublas stuff)
Now we only include a forward declaration header from ublas in the
header files. Compile time seems to be slightly reduced.

We have to clean up further and remove every use of ublas specific stuff
in the header. This will enable the end-user not to install boost to
just use Siconos.