[Events] Rework the way time is managed in Simulation/EventsManager
Now the TimeDiscretisation used in an Event is "stored" in it, and the
TimeDiscretisation is now stateless: the position in time is now saved
in the EventsManager. The Simulation ask the EventsManager for the
current time Tk, which is the time of the first Event.

Changes were required in the Examples.

A new type of TimeDiscretisation based on gmp numbers has been added.

Conflicts:
	Kernel/src/utils/SiconosAlgebra/SimpleMatrixSolvers.cpp