fix shared_ptr cycle, that lead to memleaks
Any OneStepIntegrator or OneStepNSProblem as a link to the Simulation as
member. The Simulation has maps of shared_ptr to both OneStepIntegrator
ans OneStepNSProblem. Hence there is a cycle and the memory is never
freed.
The solution adopted here is to clear the maps in the Model destructor,
thus breaking the cycle at the very end.

The real fix is to remove the cycle. This could be done by removing the
simulationLink in the OneStepIntegrator and OneStepNSProblem.