[cmake] add new DEV_MODE CMake variable (ON for dev !!)
Since the code is now on github, everyone can build the code in its current state.
This means that we cannot hide anymore the tests that we are failing because a particular feature is a WIP.
Now we can use this DEV_MODE Cmake variable to conditionnally perform tests or other action.
For instance, we can silence some warnings and make doxygen less verbose when we are not in DEV_MODE ...

This option is set to OFF in Build/CMakeLists.txt but to ON in SiconosProject. This is motivated by the fact
that we should tell people to use the meta-build procedure to build Siconos. On the other hand, the developers usually wants to
build each module separetely. The developers must have this ON :)