[Numerics,Kernel] fix issues with msan
Our drawback of using msan is that all the codepath should be instrumented for the method to work.
However, this is not possible, for instance with fortran code. One legitimate case of not initializing variable explicitly in the code, is when we use cblas_dcopy to do it.
Msan is not able to see what is done inside this function and we endup having a false-positive.
The function cblas_dcopy_msan performs a dummy initialization to 0 only if the msan sanitizer is detected and calls the function cblas_dcopy after.