[kernel/py] rework swig typemaps to use fragment
The goal was to reduce the amount of bloat code due to the typemap for
the vectors and matrices. It looks like the size of the generated code
was reduced by at least 10%, but I was expecting more

The code was also cleanup a little bit.

There is still an issue with references (+ const), that's why in the
pyRegister we have typemaps that are already in KernelTypes.i
But without them, it (randomly) fails