[mechanics proposed] Change to weak_ptr for handler pointer, merge with PImpl idiom.
Added a PImpl idiom to the BulletBroadphase class in order to further hide the
Bullet-specific data structures, but ended up merging this with the
BulletShapeHandler class since they performed essentially the same function of
holding broadphase-global data.

After some messing with valgrind to find memory leaks, changed the "handler"
pointer in SiconosShape to a std11::weak_ptr in order to avoid reference cycles.

(Alternatively a different handler with its own data could be provided
per-shape, in which case a weak_ptr would not suffice.  To be addressed?)