[numerics] scale tolerance checking frequency by adaptive doubling and halving
Just an idea, instead of having to set the error checking frequency,
if iparam[9] is non-zero, then it will double the given error checking
interval if unsuccessful, and halve it when successful.  This tends to
allow easy problems to be solved after only a few iterations, while
skipping frequent error checking for large problems.