[mechanics proposed] Add SiconosContactorSet and SiconosContactorBase
This changes the role of SiconosContactor -- no longer a collection of
shapes, but is a single shape extended to contain a reference to an
offset and collision group.

Multiple contactors are grouped by SiconosContactorSet, and the
SiconosCollisionManager can have a list of these associated with a
base position.

SiconosShape, SiconosContactor, and SiconosContactorSet can all be
shared between multiple instances.  (ie. a BodyDS can refer to a
single SiconosContactorSet pointing to the same SiconosContactor
multiple times, and the CollisionManager also maintains a set of
static SiconosContactorSets which can be added multiple times at
different positions.)

SiconosContactorBase generalises the idea that a SiconosContactorSet
can have a base position.  For static contactors this must be provided
explicitly, but for dynamic bodies the associated BodyDS is the
SiconosContactorBase.